# apps/server-nest/openapi.yaml
openapi: 3.0.0
paths:
...
  /graph/objects:
    post:
      operationId: GraphObjectsController_create
      summary: Create a graph object (initial version)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGraphObjectDto'
      responses:
        '201':
          description: Created
        '400':
          description: Validation error
      tags: &ref_21
        - Graph
      security:
        - bearer: []
      x-required-scopes:
        - graph:write
  /graph/objects/search:
    get:
      operationId: GraphObjectsController_searchObjects
      summary: Search graph objects (basic filters)
      description: >-
        Supports pagination via created_at cursor. Optional order param (asc|desc) controls chronological direction
        (default asc). Desc returns newest first; cursor always set to last item created_at in current page.
      parameters:
        - name: type
          required: true
          in: query
          schema:
            type: string
        - name: key
          required: true
          in: query
          schema:
            type: string
        - name: label
          required: true
          in: query
          schema:
            type: string
        - name: limit
          required: true
          in: query
          schema:
            type: string
        - name: cursor
          required: true
          in: query
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: Chronological direction (asc=oldest→newest, desc=newest→oldest). Default asc.
          schema:
            enum:
              - asc
              - desc
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
  /graph/objects/fts:
    get:
      operationId: GraphObjectsController_ftsSearch
      summary: Full-text search graph objects
      description: >-
        Websearch syntax over inline-populated tsvector (type, key, properties JSON). Returns newest heads ranked by
        ts_rank. Supports optional type, label, branch_id filters. Limit max 100.
      parameters:
        - name: q
          required: true
          in: query
          description: Search query (websearch syntax)
          schema:
            type: string
        - name: limit
          required: true
          in: query
          schema:
            type: string
        - name: type
          required: true
          in: query
          schema:
            type: string
        - name: label
          required: true
          in: query
          schema:
            type: string
        - name: branch_id
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
  /graph/objects/{id}:
    get:
      operationId: GraphObjectsController_get
      summary: Get latest version of a graph object
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
        '404':
          description: Not found
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
    patch:
      operationId: GraphObjectsController_patch
      summary: Patch (create new version) of a graph object
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchGraphObjectDto'
      responses:
        '200':
          description: Version created
        '400':
          description: No effective change
        '404':
          description: Not found
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:write
    delete:
      operationId: GraphObjectsController_deleteObject
      summary: Soft delete (tombstone) an object (creates new version with deleted_at)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:write
  /graph/objects/{id}/restore:
    post:
      operationId: GraphObjectsController_restoreObject
      summary: Restore a soft-deleted object (creates new version clearing deleted_at)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Restored (new version created)
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:write
  /graph/objects/{id}/history:
    get:
      operationId: GraphObjectsController_history
      summary: List version history for a graph object
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: true
          in: query
          schema:
            type: string
        - name: cursor
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectHistoryResponseDto'
        '404':
          description: Not found
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
  /graph/relationships:
    post:
      operationId: GraphObjectsController_createRel
      summary: Create a relationship (initial version) or new version if properties changed
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGraphRelationshipDto'
      responses:
        '201':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:write
  /graph/relationships/{id}:
    get:
      operationId: GraphObjectsController_getRel
      summary: Get relationship by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
    patch:
      operationId: GraphObjectsController_patchRel
      summary: Patch (create new version) of a relationship (only head version allowed)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchGraphRelationshipDto'
      responses:
        '200':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:write
    delete:
      operationId: GraphObjectsController_deleteRel
      summary: Soft delete a relationship (tombstone version)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:write
  /graph/relationships/search:
    get:
      operationId: GraphObjectsController_searchRels
      summary: Search relationships (basic filters)
      description: >-
        Supports pagination via created_at cursor. Optional order param (asc|desc) for chronological direction (default
        asc).
      parameters:
        - name: type
          required: true
          in: query
          schema:
            type: string
        - name: src_id
          required: true
          in: query
          schema:
            type: string
        - name: dst_id
          required: true
          in: query
          schema:
            type: string
        - name: limit
          required: true
          in: query
          schema:
            type: string
        - name: cursor
          required: true
          in: query
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: Chronological direction (asc=oldest→newest, desc=newest→oldest). Default asc.
          schema:
            enum:
              - asc
              - desc
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
  /graph/relationships/{id}/restore:
    post:
      operationId: GraphObjectsController_restoreRel
      summary: Restore a soft-deleted relationship
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Restored (new version created)
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:write
  /graph/relationships/{id}/history:
    get:
      operationId: GraphObjectsController_historyRel
      summary: List version history for a relationship
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: true
          in: query
          schema:
            type: string
        - name: cursor
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipHistoryResponseDto'
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
  /graph/objects/{id}/edges:
    get:
      operationId: GraphObjectsController_edges
      summary: List relationships adjacent to an object
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: direction
          required: true
          in: query
          schema:
            type: string
        - name: limit
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
  /graph/traverse:
    post:
      operationId: GraphObjectsController_traverse
      summary: Traverse the graph from one or more root object ids (bounded BFS)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TraverseGraphDto'
      responses:
        '200':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
  /graph/expand:
    post:
      operationId: GraphObjectsController_expand
      summary: Expand graph (single-pass bounded traversal with projection & edge property option)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphExpandDto'
      responses:
        '200':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
  /graph/objects/vector-search:
    post:
      operationId: GraphObjectsController_vectorSearch
      summary: Vector similarity search
      description: >-
        Returns nearest neighbors (cosine distance) for provided query vector over embedding_vec. Threshold: use
        maxDistance (preferred) or legacy minScore (acts as maximum distance). When both are supplied, maxDistance takes
        precedence.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VectorSearchDto'
      responses:
        '200':
          description: ''
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
...
  /graph/embedding-policies:
    post:
      operationId: GraphObjectsController_createEmbeddingPolicy
      summary: Create an embedding policy for a project and object type
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmbeddingPolicyDto'
      responses:
        '201':
          description: Policy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingPolicyResponseDto'
        '400':
          description: Validation error or duplicate policy
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:write
    get:
      operationId: GraphObjectsController_listEmbeddingPolicies
      summary: List all embedding policies for a project
      parameters:
        - name: project_id
          required: true
          in: query
          description: Filter policies by project ID
          schema:
            type: string
        - name: object_type
          required: false
          in: query
          description: Filter policies by object type
          schema:
            type: string
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmbeddingPolicyResponseDto'
      tags: *ref_21
      security:
        - bearer: []
      x-required-scopes:
        - graph:read
...
info:
  title: Spec Server API
  description: >-
    Generated OpenAPI spec (code-first). All protected operations include x-required-scopes and 403 examples with
    missing_scopes.
  version: 0.2.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateGraphObjectDto:
      type: object
      properties: {}
    PatchGraphObjectDto:
      type: object
      properties: {}
...
    ObjectHistoryResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VersionedObjectDto'
        next_cursor:
          type: string
          description: Present when more pages available; pass as cursor query param
      required:
        - items
    CreateGraphRelationshipDto:
      type: object
      properties: {}
    PatchGraphRelationshipDto:
      type: object
      properties: {}
...
    RelationshipHistoryResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VersionedRelationshipDto'
        next_cursor:
          type: string
          description: Present when more pages available; pass as cursor query param
      required:
        - items
    TraverseGraphDto:
      type: object
      properties: {}
    GraphExpandDto:
      type: object
      properties: {}
...